Feature: fixed the variance and std problem by changing division to a float

+++ src/Calculator/Division.py
@@ -1,2 +1,2 @@
 def divide(a,b):
-    return int(a) / int(b)
+    return float(a) / float(b)
